Completing quest automation. Split into files
