+++ Content.Server/_Stalker/AI/AINPCSystem.History.cs
@@ -1,0 +1,32 @@
+using Content.Server.Chat.Systems;
+using Content.Shared.Hands.EntitySystems;
+using Robust.Shared.GameObjects;
+using Robust.Shared.IoC;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.Json;
+using System.Text.Json.Nodes;
+using Robust.Shared.Log;
+using Content.Shared.Chat;
+using Robust.Shared.Timing;
+using System.Threading.Tasks;
+using System.Threading;
+using static Content.Server._Stalker.AI.AIManager;
+using Content.Shared.Hands.Components;
+using Robust.Shared.Player;
+using Content.Server._Stalker.AI;
+using Content.Shared._Stalker.AI;
+using Robust.Shared.Prototypes;
+using Content.Shared.Damage;
+using Content.Shared.NPC.Systems;
+using Content.Shared.Whitelist;
+using Robust.Server.Audio;
+using Content.Shared.NPC.Prototypes;
+
+namespace Content.Server._Stalker.AI
+{
+    public sealed partial class AINPCSystem : SharedAiNpcSystem
+    {
+ 
+    }
+}
