+++ Content.Server/_Stalker/AI/AINPCSystem.cs
@@ -2,6 +2,7 @@ using Content.Server.Chat.Systems;
 using Content.Shared.Hands.EntitySystems;
 using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
@@ -25,7 +26,7 @@ using Content.Shared.NPC.Prototypes;
 
 namespace Content.Server._Stalker.AI
 {
-    public sealed class AINPCSystem : SharedAiNpcSystem
+    public sealed partial class AINPCSystem : SharedAiNpcSystem
     {
         [Dependency] private readonly AIManager _aiManager = default!;
         [Dependency] private readonly ChatSystem _chatSystem = default!;
@@ -45,6 +46,7 @@ namespace Content.Server._Stalker.AI
 
             SubscribeLocalEvent<EntitySpokeEvent>(OnEntitySpoke);
             SubscribeLocalEvent<ProcessAIResponseEvent>(HandleAIResponse);
+            SubscribeLocalEvent<AiNpcComponent, ComponentShutdown>(OnComponentRemoved); // Added subscription
 
             _sawmill.Info("AI NPC System Initialized");
         }
@@ -233,7 +235,7 @@ namespace Content.Server._Stalker.AI
                         if (!component.CanChat)
                         {
                             resultMessage = "TryChat tool is disabled for this NPC.";
-                            break; // Skip execution
+                            break;
                         }
                         if (TryGetStringArgument(toolCall.Arguments, "message", out var message))
                         {
@@ -243,65 +245,84 @@ namespace Content.Server._Stalker.AI
                         else resultMessage = "Missing or invalid 'message' argument for TryChat.";
                         break;
 
-                    case nameof(TryGiveItem): // Use nameof
+                    case nameof(TryGiveItem):
                         if (!component.CanGiveItems)
                         {
                             resultMessage = "TryGiveItem tool is disabled for this NPC.";
-                            break; // Skip execution
+                            break;
                         }
-                        if (TryGetStringArgument(toolCall.Arguments, "targetPlayer", out var targetPlayer) &&
+                        if (TryGetStringArgument(toolCall.Arguments, "ckey", out var targetPlayer) &&
                             TryGetStringArgument(toolCall.Arguments, "itemPrototypeId", out var itemPrototypeId))
                         {
                             TryGetIntArgument(toolCall.Arguments, "quantity", out var quantity);
                             quantity = Math.Max(1, quantity);
-                            success = TryGiveItem(uid, targetPlayer, itemPrototypeId, quantity, npcResponse); // Pass npcResponse
+                            success = TryGiveItem(uid, targetPlayer, itemPrototypeId, quantity, npcResponse);
                             resultMessage = success ? "Give item action performed." : "Give item action failed.";
                         }
                         else resultMessage = "Missing or invalid arguments for TryGiveItem (expected targetPlayer, itemPrototypeId).";
                         break;
 
-                    case nameof(TryTakeItem): // Use nameof
+                    case nameof(TryTakeItem):
                         if (!component.CanTakeItems)
                         {
                             resultMessage = "TryTakeItem tool is disabled for this NPC.";
-                            break; // Skip execution
+                            break;
                         }
-                        if (TryGetStringArgument(toolCall.Arguments, "targetPlayer", out targetPlayer) &&
+                        if (TryGetStringArgument(toolCall.Arguments, "ckey", out targetPlayer) &&
                            TryGetStringArgument(toolCall.Arguments, "requestedItemName", out var requestedItemName))
                         {
-                            success = TryTakeItem(uid, targetPlayer, requestedItemName, npcResponse); // Pass npcResponse
+                            success = TryTakeItem(uid, targetPlayer, requestedItemName, npcResponse);
                             resultMessage = success ? "Take item action performed." : "Take item action failed.";
                         }
                         else resultMessage = "Missing or invalid arguments for TryTakeItem (expected targetPlayer, requestedItemName).";
                         break;
 
-                    case nameof(TryPunishPlayer): // Added case for the new tool
+                    case nameof(TryPunishPlayer):
                         if (!component.CanPunish)
                         {
                             resultMessage = "TryPunishPlayer tool is disabled for this NPC.";
-                            break; // Skip execution
+                            break;
                         }
-                        if (TryGetStringArgument(toolCall.Arguments, "targetPlayer", out targetPlayer) &&
-                            TryGetStringArgument(toolCall.Arguments, "reason", out var reason)) // Added reason argument
+                        if (TryGetStringArgument(toolCall.Arguments, "ckey", out targetPlayer) &&
+                            TryGetStringArgument(toolCall.Arguments, "reason", out var reason))
                         {
-                            success = TryPunishPlayer(uid, component, targetPlayer, reason, npcResponse); // Pass npcResponse
+                            success = TryPunishPlayer(uid, component, targetPlayer, reason, npcResponse);
                             resultMessage = success ? "Punish player action performed." : "Punish player action failed.";
                         }
                         else resultMessage = "Missing or invalid arguments for TryPunishPlayer (expected targetPlayer, reason).";
                         break;
 
-                    case nameof(TryOfferQuest): // Added case for the new quest tool
+                    case nameof(TryOfferQuest):
                         if (!component.CanOfferQuests)
                         {
                             resultMessage = "TryOfferQuest tool is disabled for this NPC.";
-                            break; // Skip execution
+                            break;
                         }
-                        if (TryGetStringArgument(toolCall.Arguments, "targetPlayer", out targetPlayer))
+                        if (TryGetStringArgument(toolCall.Arguments, "ckey", out targetPlayer))
                         {
-                            success = TryOfferQuest(uid, component, targetPlayer, npcResponse); // Pass npcResponse
+                            success = TryOfferQuest(uid, component, targetPlayer, npcResponse);
                             resultMessage = success ? "Offer quest action performed." : "Offer quest action failed.";
                         }
                         else resultMessage = "Missing or invalid 'targetPlayer' argument for TryOfferQuest.";
+                        break;
+
+                    case nameof(TryCompleteQuest): // <-- New Case
+                        // Check if NPC can both take and give items (required for quest completion)
+                        if (!component.CanOfferQuests)
+                        {
+                            resultMessage = "TryCompleteQuest tool requires canOfferQuests to be enabled.";
+                            break;
+                        }
+                        if (TryGetStringArgument(toolCall.Arguments, "ckey", out targetPlayer) &&
+                            TryGetStringArgument(toolCall.Arguments, "questItemId", out var questItemId) &&
+                            TryGetStringArgument(toolCall.Arguments, "rewardItemId", out var rewardItemId))
+                        {
+                            TryGetIntArgument(toolCall.Arguments, "rewardQuantity", out var rewardQuantity);
+                            rewardQuantity = Math.Max(1, rewardQuantity); // Ensure at least 1 reward
+                            success = TryCompleteQuest(uid, component, targetPlayer, questItemId, rewardItemId, rewardQuantity, npcResponse);
+                            resultMessage = success ? "Complete quest action performed." : "Complete quest action failed.";
+                        }
+                        else resultMessage = "Missing or invalid arguments for TryCompleteQuest (expected targetPlayer, questItemId, rewardItemId).";
                         break;
 
                     default:
@@ -398,8 +419,18 @@ namespace Content.Server._Stalker.AI
                 }
             }
 
-            // Basic history addition using the sanitized name
-            history.Add(new OpenRouterMessage { Role = role, Content = content, Name = sanitizedName, CKey = speakerCKey, ToolCalls = toolCalls, ToolCallId = toolCallId });
+            // Prepend CKey to user message content if available
+            string? finalContent = content;
+            if (role == "user" && !string.IsNullOrWhiteSpace(speakerCKey) && !string.IsNullOrWhiteSpace(content))
+            {
+                finalContent = $"[CKEY: {speakerCKey}] {content}";
+                _sawmill.Debug($"Prepended CKey to user message for {speakerCKey}. New content preview: {finalContent.Substring(0, Math.Min(finalContent.Length, 50))}...");
+            }
+
+            // Basic history addition using the sanitized name and potentially modified content
+            // Note: We still pass speakerCKey to the CKey property for potential internal use or future API changes,
+            // even though the current OpenRouter API might ignore it. The primary mechanism is now the content.
+            history.Add(new OpenRouterMessage { Role = role, Content = finalContent, Name = sanitizedName, ToolCalls = toolCalls, ToolCallId = toolCallId });
 
             // Trimming is handled by TrimHistory calls before this method.
         }
@@ -489,6 +520,13 @@ namespace Content.Server._Stalker.AI
             if (component.CanTakeItems)
                 descriptions.Add(GetTakeItemToolDescription());
 
+            // Add CompleteQuest tool if NPC can both take and give items
+            // AND has both quest items and givable items defined (as rewards)
+            if (component.CanOfferQuests && component.QuestItems.Count > 0 && component.GivableItems.Count > 0)
+            {
+                descriptions.Add(GetCompleteQuestToolDescription(component)); // Needs component for item lists
+            }
+
             // Add the punish tool description if the component allows it
             // AND the tool is enabled
             if (component.CanPunish && (component.PunishmentDamage != null || component.PunishmentSound != null))
@@ -497,466 +535,6 @@ namespace Content.Server._Stalker.AI
             }
             return descriptions;
         }
-
-        private string GetChatToolDescription() // Removed component parameter
-        {
-            // Simplified description
-            var description = "Respond verbally to a user or initiate conversation.";
-
-            return $@"{{
-                ""name"": ""TryChat"",
-                ""description"": ""{JsonEncodedText.Encode(description)}"",
-                ""parameters"": {{
-                    ""type"": ""object"",
-                    ""properties"": {{
-                        ""message"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The primary message the NPC should say.""
-                        }},
-                        ""npcResponse"": {{
-                            ""type"": ""string"",
-                            ""description"": ""Optional alternative message the NPC says. If provided, this is spoken instead of 'message'.""
-                        }}
-                    }},
-                    ""required"": [""message""]
-                }}
-            }}";
-        }
-
-
-        private string GetGiveItemToolDescription(AiNpcComponent component)
-        {
-            // Build the list of allowed items for the description from GivableItems
-            var allowedItemsList = component.GivableItems
-                .Select(item => $"- {item.ProtoId} (Max: {item.MaxQuantity}, Rarity: {item.Rarity})")
-                .ToList();
-            var allowedItemsString = allowedItemsList.Count > 0
-                ? string.Join("\n", allowedItemsList)
-                : "None";
-
-            // Simplified description, emphasizing it's for rewards/gifts
-            var description = $@"Give a specified item (as a reward or gift) to a player. You can only give items from this list:
-                {allowedItemsString}";
-
-            return $@"{{
-                ""name"": ""TryGiveItem"",
-                ""description"": ""{JsonEncodedText.Encode(description)}"",
-                ""parameters"": {{
-                    ""type"": ""object"",
-                    ""properties"": {{
-                         ""targetPlayer"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The CKey (username) of the player to give the item to.""
-                        }},
-                        ""itemPrototypeId"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The exact prototype ID of the item from the allowed list.""
-                        }},
-                        ""quantity"": {{
-                            ""type"": ""integer"",
-                            ""description"": ""Number of items to give (defaults to 1, respects max quantity)."",
-                            ""default"": 1
-                        }},
-                        ""npcResponse"": {{
-                            ""type"": ""string"",
-                            ""description"": ""Optional message the NPC says while giving the item (e.g., 'Here's your reward.').""
-                        }}
-                    }},
-                    ""required"": [""targetPlayer"", ""itemPrototypeId""]
-                }}
-            }}";
-        }
-
-        private string GetOfferQuestToolDescription(AiNpcComponent component)
-        {
-            // Build the list of possible quest items
-            var questItemsList = component.QuestItems
-                .Select(item => $"- {item.ProtoId} (Rarity: {item.Rarity})")
-                .ToList();
-            var questItemsString = questItemsList.Count > 0
-                ? string.Join("\n", questItemsList)
-                : "None";
-
-            // Description for the new quest tool
-            var description = $@"Offer a quest to a player to retrieve ONE specific item. Checks if the player already has an active quest from you. Possible items to request:
-                {questItemsString}";
-
-            return $@"{{
-                ""name"": ""TryOfferQuest"",
-                ""description"": ""{JsonEncodedText.Encode(description)}"",
-                ""parameters"": {{
-                    ""type"": ""object"",
-                    ""properties"": {{
-                         ""targetPlayer"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The CKey (username) of the player to offer the quest to.""
-                        }},
-                        ""npcResponse"": {{
-                            ""type"": ""string"",
-                            ""description"": ""REQUIRED message the NPC says when offering the quest (e.g., 'Hey stalker, bring me a BoarHoof and I'll pay you.').""
-                        }}
-                    }},
-                    ""required"": [""targetPlayer"", ""npcResponse""]
-                }}
-            }}";
-        }
-
-
-        private string GetTakeItemToolDescription()
-        {
-            // Simplified description, emphasizing it's for quest items/specific requests
-            var description = "Request a specific item (e.g., a quest item) from a player and attempt to take it if they hold it out.";
-            return $@"{{
-                ""name"": ""TryTakeItem"",
-                ""description"": ""{JsonEncodedText.Encode(description)}"",
-                ""parameters"": {{
-                    ""type"": ""object"",
-                    ""properties"": {{
-                         ""targetPlayer"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The CKey (username) of the player to request the item from.""
-                        }},
-                        ""requestedItemName"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The exact prototype ID of the item the NPC wants to receive (e.g., 'MutantPartBoarHoof').""
-                        }},
-                        ""npcResponse"": {{
-                            ""type"": ""string"",
-                            ""description"": ""REQUIRED message the NPC says while requesting/taking the item (e.g., 'Alright, let's see that BoarHoof. Hold it out.').""
-                        }}
-                    }},
-                    ""required"": [""targetPlayer"", ""requestedItemName"", ""npcResponse""]
-                }}
-            }}";
-        }
-
-        private string GetPunishPlayerToolDescription() // Removed component parameter
-        {
-            // Simplified description
-            var description = "Punish a player perceived as rude, lying, or hostile by attacking them. Use sparingly.";
-
-            return $@"{{
-                ""name"": ""TryPunishPlayer"",
-                ""description"": ""{JsonEncodedText.Encode(description)}"",
-                ""parameters"": {{
-                    ""type"": ""object"",
-                    ""properties"": {{
-                         ""targetPlayer"": {{
-                            ""type"": ""string"",
-                            ""description"": ""The CKey (username) of the player to punish.""
-                        }},
-                        ""reason"": {{
-                            ""type"": ""string"",
-                            ""description"": ""A brief reason for the punishment (e.g., 'Insults', 'Attempted scam').""
-                        }},
-                        ""npcResponse"": {{
-                            ""type"": ""string"",
-                            ""description"": ""Optional message the NPC shouts while punishing (e.g., 'You asked for it!').""
-                        }}
-                    }},
-                    ""required"": [""targetPlayer"", ""reason""]
-                }}
-            }}";
-        }
-
-
-        // Actual tool methods, invoked locally after AI response parsing
-
-        /// <summary>
-        /// Makes the NPC speak. Handles the npcResponse parameter to avoid double-speaking if called from ExecuteToolCall.
-        /// </summary>
-        public bool TryChat(EntityUid npc, string message, string? npcResponse = null)
-        {
-            // Prioritize npcResponse if provided, otherwise use message.
-            var messageToSpeak = !string.IsNullOrWhiteSpace(npcResponse) ? npcResponse : message;
-
-            if (string.IsNullOrWhiteSpace(messageToSpeak))
-                return false;
-
-            // Avoid logging the same message twice if npcResponse was handled in ExecuteToolCall
-            if (messageToSpeak != npcResponse)
-                _sawmill.Debug($"NPC {ToPrettyString(npc)} executing chat: {messageToSpeak}");
-
-            _chatSystem.TrySendInGameICMessage(npc, messageToSpeak, InGameICChatType.Speak, hideChat: false);
-            return true;
-        }
-
-
-        // Updated TryGiveItem to include npcResponse
-        public bool TryGiveItem(EntityUid npc, string targetPlayerIdentifier, string itemPrototypeId, int quantity, string? npcResponse = null)
-        {
-            // npcResponse is handled by ExecuteToolCall before this method is run.
-            _sawmill.Debug($"NPC {ToPrettyString(npc)} attempting give item: Proto='{itemPrototypeId}', Qty={quantity}, Target='{targetPlayerIdentifier}'");
-
-            // --- 0. Get NPC Component ---
-            if (!TryComp<AiNpcComponent>(npc, out var aiComp))
-            {
-                _sawmill.Error($"NPC {ToPrettyString(npc)} is missing AiNpcComponent in TryGiveItem.");
-                return false;
-            }
-
-            // --- 1. Find Target Player ---
-            EntityUid? targetPlayer = FindPlayerByIdentifier(targetPlayerIdentifier);
-            if (targetPlayer == null || !targetPlayer.Value.Valid)
-            {
-                _sawmill.Warning($"Could not find target player '{targetPlayerIdentifier}' for TryGiveItem.");
-                // Maybe chat failure? TryChat(npc, $"Who is {targetPlayerIdentifier}?");
-                return false;
-            }
-
-            // --- 2. Validate Item Against Givable List & Quantity ---
-            var givableItemInfo = aiComp.GivableItems.FirstOrDefault(item => item.ProtoId.Id.Equals(itemPrototypeId, StringComparison.OrdinalIgnoreCase));
-
-            if (givableItemInfo == null)
-            {
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} tried to give non-givable item '{itemPrototypeId}'. Denying.");
-                // TryChat(npc, $"I can't give you one of those."); // Feedback
-                return false;
-            }
-
-            if (quantity <= 0)
-            {
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} tried to give zero or negative quantity ({quantity}) of '{itemPrototypeId}'. Setting to 1.");
-                quantity = 1;
-            }
-
-            if (quantity > givableItemInfo.MaxQuantity)
-            {
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} tried to give {quantity} of '{itemPrototypeId}', but max allowed is {givableItemInfo.MaxQuantity}. Clamping quantity.");
-                quantity = givableItemInfo.MaxQuantity;
-                // TryChat(npc, $"Whoa there, I can only give you {quantity} of those."); // Feedback
-            }
-
-            // --- 3. Validate Prototype ID ---
-            if (!_proto.HasIndex<EntityPrototype>(itemPrototypeId))
-            {
-                _sawmill.Warning($"Invalid prototype ID '{itemPrototypeId}' requested by NPC {ToPrettyString(npc)} for TryGiveItem.");
-                return false;
-            }
-
-            // --- 4. Check Range & Interaction ---
-            if (!Transform(npc).Coordinates.TryDistance(EntityManager, Transform(targetPlayer.Value).Coordinates, out var distance) || distance > 2.0f)
-            {
-                _sawmill.Warning($"Target player {ToPrettyString(targetPlayer.Value)} too far for NPC {ToPrettyString(npc)} to give item.");
-                // TryChat(npc, $"Get closer if you want this."); // Feedback
-                return false;
-            }
-
-            // --- 5. Spawn and Give Items ---
-            var npcCoords = Transform(npc).Coordinates;
-            if (!npcCoords.IsValid(EntityManager))
-            {
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} has invalid coordinates, cannot spawn items.");
-                return false;
-            }
-
-            int givenCount = 0;
-            for (int i = 0; i < quantity; i++)
-            {
-                var spawnedItem = Spawn(itemPrototypeId, npcCoords);
-                if (!_hands.TryPickupAnyHand(targetPlayer.Value, spawnedItem, checkActionBlocker: false))
-                {
-                    _sawmill.Warning($"Failed direct pickup for item {i + 1}/{quantity} ({ToPrettyString(spawnedItem)}) by {ToPrettyString(targetPlayer.Value)}. Dropping near NPC.");
-                    Transform(spawnedItem).Coordinates = npcCoords;
-                    // If dropping near, maybe don't count as "given"? Or maybe do? Let's count it for now.
-                    givenCount++; // Count even if dropped nearby
-                }
-                else
-                {
-                    givenCount++;
-                }
-            }
-
-            if (givenCount > 0)
-                _sawmill.Info($"NPC {ToPrettyString(npc)} successfully gave {givenCount}/{quantity} of '{itemPrototypeId}' to {ToPrettyString(targetPlayer.Value)}.");
-            else
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} failed to give any '{itemPrototypeId}' to {ToPrettyString(targetPlayer.Value)} (pickup/drop failed).");
-
-            return givenCount > 0;
-        }
-
-        /// <summary>
-        /// Attempts to take a specified item from a player's hands.
-        /// </summary>
-        public bool TryTakeItem(EntityUid npc, string targetPlayerIdentifier, string requestedItemName, string? npcResponse = null)
-        {
-            // npcResponse is handled by ExecuteToolCall.
-            _sawmill.Debug($"NPC {ToPrettyString(npc)} attempting take item: ItemName='{requestedItemName}', Target='{targetPlayerIdentifier}'");
-
-            // --- 1. Find Target Player ---
-            EntityUid? targetPlayer = FindPlayerByIdentifier(targetPlayerIdentifier);
-            if (targetPlayer == null || !targetPlayer.Value.Valid)
-            {
-                _sawmill.Warning($"Could not find target player '{targetPlayerIdentifier}' for TryTakeItem.");
-                // TryChat(npc, $"Who's {targetPlayerIdentifier}?");
-                return false;
-            }
-
-            // --- 2. Find Item in Player's Active Hand ---
-            // We only check the *active* hand for simplicity now.
-            EntityUid? itemToTake = null;
-            if (_hands.TryGetActiveItem(targetPlayer.Value, out var activeItem))
-            {
-                var proto = Prototype(activeItem.Value);
-                if (proto != null && string.Equals(proto.ID, requestedItemName, StringComparison.OrdinalIgnoreCase))
-                {
-                    itemToTake = activeItem;
-                }
-            }
-
-            if (itemToTake == null || !itemToTake.Value.Valid)
-            {
-                _sawmill.Warning($"Player {ToPrettyString(targetPlayer.Value)} does not have '{requestedItemName}' in active hand for TryTakeItem.");
-                // TryChat(npc, $"You don't seem to be holding a {requestedItemName}. Hold it out."); // Feedback
-                return false;
-            }
-
-            // --- 3. Check Range & Interaction ---
-            if (!Transform(npc).Coordinates.TryDistance(EntityManager, Transform(targetPlayer.Value).Coordinates, out var distance) || distance > 2.0f)
-            {
-                _sawmill.Warning($"Target player {ToPrettyString(targetPlayer.Value)} too far for NPC {ToPrettyString(npc)} to take item.");
-                // TryChat(npc, $"Bring that {requestedItemName} closer."); // Feedback
-                return false;
-            }
-
-            // --- 4. Perform Transfer (Original drop + move logic) ---
-            // Player drops the item at their feet first.
-            // A real implementation might need a more robust interaction system (e.g., trade window).
-            if (_hands.TryDrop(targetPlayer.Value, itemToTake.Value, checkActionBlocker: false)) // Player drops the item
-            {
-                // Now move the dropped item to the NPC's location
-                var npcTransform = Transform(npc);
-                var itemTransform = Transform(itemToTake.Value);
-                itemTransform.Coordinates = npcTransform.Coordinates; // Move item to NPC's feet
-
-                _sawmill.Info($"NPC {ToPrettyString(npc)} successfully received item {ToPrettyString(itemToTake.Value)} from {ToPrettyString(targetPlayer.Value)} (dropped at NPC location).");
-                return true; // Success is player dropping the item near the NPC
-            }
-            else
-            {
-                _sawmill.Warning($"Player {ToPrettyString(targetPlayer.Value)} failed to drop item {ToPrettyString(itemToTake.Value)} for NPC {ToPrettyString(npc)}.");
-                // TryChat(npc, $"Looks like you couldn't drop that {requestedItemName}. Try again?"); // Feedback
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Attempts to punish a player by applying damage.
-        /// </summary>
-        public bool TryPunishPlayer(EntityUid npc, AiNpcComponent aiComp, string targetPlayerIdentifier, string reason, string? npcResponse = null)
-        {
-            // npcResponse is handled by ExecuteToolCall.
-            _sawmill.Debug($"NPC {ToPrettyString(npc)} attempting to punish player: Target='{targetPlayerIdentifier}', Reason='{reason}'");
-
-            // --- 1. Find Target Player ---
-            EntityUid? targetPlayer = FindPlayerByIdentifier(targetPlayerIdentifier);
-            if (targetPlayer == null || !targetPlayer.Value.Valid)
-            {
-                _sawmill.Warning($"Could not find target player '{targetPlayerIdentifier}' for TryPunishPlayer.");
-                return false;
-            }
-
-
-            // Check if the target is whitelisted from punishment by this NPC
-            if (aiComp.PunishmentWhitelist != null && _whitelistSystem.IsWhitelistPass(aiComp.PunishmentWhitelist, targetPlayer.Value))
-            {
-                _sawmill.Info($"Target player {ToPrettyString(targetPlayer.Value)} is whitelisted. Punishment aborted.");
-                // TryChat(npc, $"I can't touch {Name(targetPlayer.Value)}."); // Feedback
-                return false;
-            }
-
-            // --- 3. Check Range ---
-            const float punishRange = 5.0f;
-            if (!Transform(npc).Coordinates.TryDistance(EntityManager, Transform(targetPlayer.Value).Coordinates, out var distance) || distance > punishRange)
-            {
-                _sawmill.Warning($"Target player {ToPrettyString(targetPlayer.Value)} too far ({distance}m) for NPC {ToPrettyString(npc)} to punish.");
-                // TryChat(npc, $"Get back here, coward!"); // Feedback
-                return false;
-            }
-
-            // --- 4. Apply Punishment ---
-            bool damageApplied = false;
-            if (aiComp.PunishmentDamage != null)
-            {
-                var damageResult = _damageable.TryChangeDamage(targetPlayer.Value, aiComp.PunishmentDamage, ignoreResistances: true);
-                damageApplied = damageResult != null; // Don't fucking touch it. There is no Total damage!
-            }
-            else
-            {
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} tried to punish, but no PunishmentDamage is defined in its AiNpcComponent.");
-            }
-
-            // Play sound effect if defined
-            if (aiComp.PunishmentSound != null)
-            {
-                _audio.PlayPvs(aiComp.PunishmentSound, Transform(npc).Coordinates);
-            }
-
-            return damageApplied;
-        }
-
-        /// <summary>
-        /// Offers a quest to the player if they don't already have one.
-        /// TODO: Implement actual quest tracking.
-        /// </summary>
-        public bool TryOfferQuest(EntityUid npc, AiNpcComponent aiComp, string targetPlayerIdentifier, string? npcResponse = null)
-        {
-            // npcResponse is handled by ExecuteToolCall. It should contain the quest offer text.
-            _sawmill.Debug($"NPC {ToPrettyString(npc)} attempting to offer quest to: Target='{targetPlayerIdentifier}'");
-
-            // --- 1. Find Target Player ---
-            EntityUid? targetPlayer = FindPlayerByIdentifier(targetPlayerIdentifier);
-            if (targetPlayer == null || !targetPlayer.Value.Valid)
-            {
-                _sawmill.Warning($"Could not find target player '{targetPlayerIdentifier}' for TryOfferQuest.");
-                // TryChat(npc, $"Who are you talking about?");
-                return false;
-            }
-
-            // --- 2. Check if Player Already Has Quest (Placeholder) ---
-            // TODO: Implement a real quest tracking system. This is just a placeholder.
-            // For now, we'll pretend they never have a quest and always offer one.
-            bool playerHasActiveQuest = false; // Replace with actual check
-            if (playerHasActiveQuest)
-            {
-                _sawmill.Info($"Player {ToPrettyString(targetPlayer.Value)} already has an active quest from {ToPrettyString(npc)}. Cannot offer another.");
-                // TryChat(npc, $"Finish the job I already gave you first!"); // Feedback
-                return false;
-            }
-
-            // --- 3. Check if NPC has Quest Items Defined ---
-            if (aiComp.QuestItems == null || aiComp.QuestItems.Count == 0)
-            {
-                _sawmill.Warning($"NPC {ToPrettyString(npc)} tried to offer quest, but has no QuestItems defined.");
-                // TryChat(npc, $"I don't have any work for you right now."); // Feedback
-                return false;
-            }
-
-            // --- 4. Select a Quest Item (Simple Random for now) ---
-            // TODO: Implement better quest selection logic (e.g., based on rarity, player level, etc.)
-            var random = new Random();
-            var questItemInfo = aiComp.QuestItems[random.Next(aiComp.QuestItems.Count)];
-            var questItemId = questItemInfo.ProtoId;
-
-            // --- 5. Record the Quest (Placeholder) ---
-            // TODO: Record that targetPlayer now has a quest for questItemId from npc.
-            _sawmill.Info($"NPC {ToPrettyString(npc)} offered quest for '{questItemId}' to {ToPrettyString(targetPlayer.Value)}. (Quest tracking not implemented)");
-
-            // --- 6. Communicate Offer ---
-            // The actual quest offer text (e.g., "Bring me a DogTail") should be in the npcResponse parameter,
-            // which was already spoken by ExecuteToolCall. We don't need to chat again here unless
-            // the npcResponse was missing (which the tool description requires).
-            if (string.IsNullOrWhiteSpace(npcResponse))
-            {
-                _sawmill.Warning($"TryOfferQuest called without npcResponse for {ToPrettyString(npc)}. Tool description requires it.");
-                // Fallback chat if npcResponse was missing
-                TryChat(npc, $"Hey {Name(targetPlayer.Value)}, I need someone to fetch me a {questItemId}. Interested?");
-            }
-
-
-            return true; // Success means the offer was made (even if tracking failed)
-        }
-
 
         // --- Placeholder Helper Methods ---
 
@@ -974,30 +552,6 @@ namespace Content.Server._Stalker.AI
                 }
             }
             _sawmill.Warning($"Could not find player with CKey: {ckeyIdentifier}");
-            return null;
-        }
-
-        // TODO: Replace with robust inventory search
-        private EntityUid? FindItemInHands(EntityUid holder, string itemName)
-        {
-            // Use TryComp to get the HandsComponent
-            if (!TryComp<HandsComponent>(holder, out var handsComp))
-                return null;
-
-            // Now enumerate hands using the component
-            foreach (var hand in _hands.EnumerateHands(holder, handsComp))
-            {
-                // Check if the hand is holding an entity
-                if (hand.HeldEntity is { } heldEntityValue)
-                {
-                    var proto = Prototype(heldEntityValue);
-
-                    if (proto != null && string.Equals(proto.ID, itemName, StringComparison.OrdinalIgnoreCase))
-                    {
-                        return heldEntityValue;
-                    }
-                }
-            }
             return null;
         }
     }
