+++ Resources/ConfigPresets/StalkerBuild/stdebug.toml
@@ -83,7 +83,7 @@ apikey = "sk-or-v1-4dc0bde460a377ddf72a7320615728b95931255a23bed9ef59aa48f51aa15
 #model = "google/gemini-2.5-flash-preview"
 #model = "openai/gpt-4.1"
 #model = "google/gemini-2.5-flash-preview:thinking"
-model = "openai/gpt-4o"
+model = "openai/gpt-4o-mini"
 url = "https://openrouter.ai/api/v1"
