+++ Resources/Prototypes/_Stalker/ScriptedEntities/Portals/mapLoader.yml
@@ -2,13 +2,13 @@
   id: StalkerBaseMaps
   mapPaths:
     RedForest: /Maps/_ST/World/RedForestV63.yml
-    KordonUnerground: /Maps/_ST/World/KordonPodzemV42.yml
-    Svalka: /Maps/_ST/World/SvalkaV147.yml
-    SvalkaUnderground: /Maps/_ST/World/SvalkaPodzemkaV52.yml
-    Boloto: /Maps/_ST/World/BolotaV27.yml
-    TemLoschina: /Maps/_ST/World/TemnajLohinaV45.yml
-    RostokAndWilds: /Maps/_ST/World/RostokAndWildsV18.yml
-    UndegroundWildsBar100rads: /Maps/_ST/World/UndegroundWildsBar100radsV33.yml
-    Topi: /Maps/_ST/World/TopiV30.yml
-    TemDolina: /Maps/_ST/World/TemnajDolinaV68.yml
-    Agroprom: /Maps/_ST/World/AgropromV74.yml
+    # KordonUnerground: /Maps/_ST/World/KordonPodzemV42.yml
+    # Svalka: /Maps/_ST/World/SvalkaV147.yml
+    # SvalkaUnderground: /Maps/_ST/World/SvalkaPodzemkaV52.yml
+    # Boloto: /Maps/_ST/World/BolotaV27.yml
+    # TemLoschina: /Maps/_ST/World/TemnajLohinaV45.yml
+    # RostokAndWilds: /Maps/_ST/World/RostokAndWildsV18.yml
+    # UndegroundWildsBar100rads: /Maps/_ST/World/UndegroundWildsBar100radsV33.yml
+    # Topi: /Maps/_ST/World/TopiV30.yml
+    # TemDolina: /Maps/_ST/World/TemnajDolinaV68.yml
+    # Agroprom: /Maps/_ST/World/AgropromV74.yml
